scripts/requirements.txt
openai>=1.0.0
pinecone-client>=3.0.0
pypdf>=3.0.0
python-markdown>=3.5.0
